+++ sources/src/sprites.c
@@ -281,7 +281,7 @@ SDL_Surface *sprites_get_door(struct sprites *sprites, enum door_status status) 
     return NULL;
 }
 
-SDL_Surface *sprites_get_bomb(struct sprites *sprites, int ttl) {
+SDL_Surface *sprites_get_bomb(struct sprites *sprites, uint8_t ttl) {
     assert(sprites->bomb_img[ttl - 1]);
     return sprites->bomb_img[ttl - 1];
 }
